Fixed HueSlider color and moved Handle2d to own file
